requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.0

